Added javadoc to classes.
This should explain what each part of the application does.  I have
decided to draw a line under it here, as the instructions are very
open ended and are open to interpretation.

If this was a production system it would most likely contain a GUI front
end of some description, a backend and a database of some kind to persist
data.  This would all need to be properly secured.  Obviously this would
take far longer than the 3-8 hours assigned to this task.